Add files via upload
This project focuses on controlling a robotic arm using hand gestures, with OpenCV employed to track and interpret hand movements. The Arduino serves as the servo control module, receiving serial data from the Python code to manipulate the arm. The PWM signals are mapped in the code according to the position of the selected segment within each video frame. The robotic arm supports three types of motion: rotation around the vertical axis, rotation around the horizontal axis, and the gripper movement. OpenCV handles the image processing, while the Python program manages the overall operation. Additionally, an automated thresholding procedure simplifies segment selection, allowing users to choose the target segment directly from a Tkinter interface. Once selected, the system tracks the segment attached to the hand with a colored marker such as green tape and moves the robotic arm in response to the user’s hand gestures.